extract io module
