+++ rlib-compiler/build.gradle
@@ -1,5 +1,6 @@
 
 dependencies {
   api projects.rlibCommon
+  api projects.rlibIo
   testImplementation projects.rlibLoggerImpl
 }
