+++ rlib-plugin-system/src/main/java/javasabr/rlib/plugin/system/impl/BasePluginSystem.java
@@ -25,11 +25,11 @@ import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.concurrent.atomic.AtomicReference;
 import javasabr.rlib.classpath.ClassPathScannerFactory;
 import javasabr.rlib.common.util.ClassUtils;
-import javasabr.rlib.common.util.FileUtils;
 import javasabr.rlib.common.util.Utils;
 import javasabr.rlib.common.util.array.Array;
 import javasabr.rlib.common.util.array.ReadOnlyArray;
 import javasabr.rlib.common.util.dictionary.ObjectDictionary;
+import javasabr.rlib.io.util.FileUtils;
 import javasabr.rlib.logger.api.Logger;
 import javasabr.rlib.logger.api.LoggerManager;
 import javasabr.rlib.plugin.system.ConfigurablePluginSystem;
